% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-nba-player.R
\name{best_nba_player}
\alias{best_nba_player}
\title{Find the best NBA players based on their salaries and score stats}
\usage{
best_nba_player(top, year)
}
\arguments{
\item{top}{Integer, input the number of top NBA players}

\item{year}{Integer, input the year to consider for score stats and salaries}
}
\value{
A dataframe of the top NBA players based on their salaries and score stats
}
\description{
Find the best NBA players based on their salaries and score stats
}
\examples{
best_nba_player(top = 4, year = 2021)
}
