% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/championshipPrediction.R
\name{championship.Predict.2023}
\alias{championship.Predict.2023}
\title{championship prediction}
\usage{
championship.Predict.2023()
}
\value{
a vector, containing the teams with the number of series that it is expected to win, in the decreasing order
}
\description{
championship prediction
}
\examples{
library(NBADP)
championship.Predict.2023()

}
