% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{correlation.analysis}
\alias{correlation.analysis}
\title{analysis of different factors' influence on the result of the game}
\usage{
correlation.analysis()
}
\value{
the coefficients of each factor, representing the weight of each factor
}
\description{
analysis of different factors' influence on the result of the game
}
\examples{
library(NBADP)
correlation.analysis()
}
