% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataPlayer.2020.2021}
\alias{dataPlayer.2020.2021}
\title{predict the MVP of this year}
\format{
a data frame with 705 rows and 31 columns
\describe{
  \item{Rk}{the order of the player according to the name}
  \item{Player}{the name of the player}
  \item{Pos}{the position of each player(PG,SG,SF,PF,C)}
  \item{Age}{the age of each player}
  \item{Tm}{the team that the player plays in}
  \item{G}{the number of games that the player plays}
  \item{GS}{the number of games that the player acts as the starter}
  \item{MP}{the average time that the player plays per game}
  \item{FGM}{the average field goals that a player makes per game}
  \item{FGA}{the average field goals that a player attempts per game}
  \item{FG\%}{the average field goal percentage of a player per game}
  \item{3p}{the average three points that a player makes per game}
  \item{3PA}{the average three points that a player attempts per game}
  \item{3P\%}{the average three point percentage of a player per game}
  \item{2p}{the average two points that a player makes per game}
  \item{2PA}{the average two points that a player attempts per game}
  \item{2P\%}{the average two point percentage of a player per game}
  \item{eFG\%}{effective field goal percentage of a player per game}
  \item{FTM}{the average free throws that a player makes per game}
  \item{FTA}{the average free throws that a player attempts per game}
  \item{FT\%}{the average free throw percentage of a player per game}
  \item{ORB}{the average offensive rebounds that a player gets per game}
  \item{DRB}{the average defensive rebounds that a player gets per game}
  \item{TRB}{the average total rebounds that a player gets per game}
  \item{AST}{the average assists that a player gets per game}
  \item{STL}{the average steals that a player gives per game}
  \item{BLK}{the average blocks that a player gives per game}
  \item{TOV}{the average turnovers that a player has per game}
  \item{PF}{the average fouls that a player has per game}
  \item{PTS}{the average points that a player gets per game}
  \item{MVP}{the number that a player gets according to the voting result. There are 15 players who get the votes.
  It is equal to (16-n) if the player's rank is n among the 15 candidates.
  For those who has no votes, it is 0.(0--15)}
}
}
\usage{
dataPlayer.2020.2021
}
\description{
A data set containing data of the regular season 2020-2021 of all players, along with the regular MVP results.
Another data set containing the regular season 2022-2023 of all players.
The variables of the first data set are as follows.
}
\keyword{datasets}
