% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSalariesRaw}
\alias{DataSalariesRaw}
\title{NBA_Salaries_1990_2023_}
\format{
A data frame with 15857 rows and 5 variables:
\describe{
  \item{...1}{variable ranking}
  \item{playerName}{player name}
  \item{seasonStartYear}{the starting year of an NBA season}
  \item{salary}{the actual salary of an NBA player}
  \item{inflationAdjSalary}{the salary adjusted for inflation}
}
}
\usage{
DataSalariesRaw
}
\description{
A dataset containing NBA players' salaries from 1990 to 2023. The variables are as follows:
}
\keyword{datasets}
