% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataScoreRaw}
\alias{DataScoreRaw}
\title{NBA_Player_Box_Score_Stats_1950_2022_}
\format{
A data frame with 25080 rows and 29 variables:
\describe{
  \item{...1}{variable ranking}
  \item{Season}{the season in which the game was played}
  \item{Game_ID}{a unique identifier for each game}
  \item{PLAYER_NAME}{the name of the player}
  \item{Team}{the team the player played for}
  \item{GAME_DATE}{the date the game was played}
  \item{MATCHUP}{the matchup of the game}
  \item{WL}{the outcome of the game for the player's team}
  \item{MIN}{the number of minutes the player played in the game}
  \item{FGM}{the number of field goals made by the player}
  \item{FGA}{the number of field goals attempted by the player}
  \item{FG_PCT}{the player's field goal percentage (FGM/FGA)}
  \item{FG3M}{the number of three-point field goals made by the player}
  \item{FG3A}{the number of three-point field goals attempted by the player}
  \item{FG3_PCT}{the player's three-point field goal percentage (FG3M/FG3A)}
  \item{FTM}{the number of free throws made by the player}
  \item{FTA}{the number of free throws attempted by the player}
  \item{FT_PCT}{the player's free throws attempted by the player}
  \item{OREB}{the number of offensive rebounds the player had in the game}
  \item{DREB}{the numbver of defensive rebounds the player had in the game}
  \item{REB}{the total number of rebounds the player had in the game}
  \item{AST}{the number of assists the player had in the game}
  \item{STL}{the number of steals the player had in the game}
  \item{BLK}{the number of blocks the player had in the game}
  \item{TOV}{the number of turnovers the player had in the game}
  \item{PF}{the number of personal fouls the player scored in the game}
  \item{PTS}{the total number of points the player scored in the game}
  \item{PLUS_MINUS}{the point differential when the player was on the court}
  \item{VIDEO_AVAILABLE}{a flag indicating whether or not video of the game is available for analysis}
}
}
\usage{
DataScoreRaw
}
\description{
A dataset collected from box scores, containing detailed statistics on individual player performances in NBA games from 1950 to 2022.
The variables are as follows:
}
\keyword{datasets}
