% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{NBA contract predict model datasets}
\format{
A data frame with 128 rows and 51 variables:
\describe{
  \item{Player}{player name}
  \item{salarya}{contract's salary in 2022-2023}
  \item{Guaranteed}{the amount of a player's remaining salary that is
   guaranteed}
  \item{Pos}{position}
  \item{Age}{player's age}
  \item{G}{games}
  \item{MP}{minutes played}
  \item{PER}{player efficiency rating: a measure of per-minute production
   standardized such that the league average is 15 (-45.2--76.2)}
  \item{TSp}{true shooting percentage: a measure of shooting efficiency that
   takes into account 2-point field goals, 3-point field goals, and free throw}
  \item{X3PAr}{3-point attempt rate: percentage of FG attempts from 3-point
   range}
  \item{FTr}{Free Throw attempt rate: number of FT attempts per FG attempt}
  \item{ORBp}{Offensive Rebound Percentage: an estimate of the percentage of
   available offensive rebounds a player grabbed while they were on the floor}
  \item{DRBp}{Defensive Rebound Percentage: an estimate of the percentage of
   available defensive rebounds a player grabbed while they were on the floor}
  \item{TRBp}{Total Rebound Percentage: an estimate of the percentage of
   available rebounds a player grabbed while they were on the floor}
  \item{ASTp}{Assist Percentage: an estimate of the percentage of teammate
   field goals a player assisted while they were on the floor}
  \item{STLp}{Steal Percentage: an estimate of the percentage of opponent
   possessions that end with a steal by the player while they were on the floor}
  \item{BLKp}{Block Percentage: an estimate of the percentage of opponent
   two-point field goal attempts blocked by the player while they were on the
   floor}
  \item{TOVp}{Turnover Percentage: an estimate of turnovers committed per 100
   plays}
  \item{USGp}{Usage Percentage: an estimate of the percentage of team plays
   used by a player while they were on the floor}
  \item{OWS}{Offensive Win Shares: an estimate of the number of wins
   contributed by a player due to offense}
  \item{DWS}{Defensive Win Shares: an estimate of the number of wins
   contributed by a player due to defense}
  \item{WS}{Win Shares: an estimate of the number of wins contributed by a
   player}
  \item{WSd48}{Win Shares Per 48 Minutes}{an estimate of the number of wins
   contributed by a player per 48 minutes (league average is approximately
   .100)}
  \item{OBPM}{Offensive Box Plus/Minus: a box score estimate of the offensive
   points per 100 possessions a player contributed above a league-average
   player, translated to an average team}
  \item{DBPM}{Defensive Box Plus/Minus: a box score estimate of the defensive
   points per 100 possessions a player contributed above a league-average
   player, translated to an average team}
  \item{BPM}{Box Plus/Minus: a box score estimate of the points per 100
   possessions a player contributed above a league-average player, translated
   to an average team}
  \item{VORP}{Value over Replacement Player: a box score estimate of the
   points per 100 Team possessions that a player contributed above a
   replacement-level (-2.0) player, translated to an average team and prorated
   to an 82-game season}
  \item{GS}{game started}
  \item{FG}{field goals per game}
  \item{FGA}{field goal attempts per game}
  \item{FGp}{field goal percentage}
  \item{X3P}{3-point field goals per game}
  \item{X3PA}{3-point field goal attempts per game}
  \item{X3Pp}{3-point field goal percentage}
  \item{X2P}{2-point field goals per game}
  \item{X2PA}{2-point field goal attempts per game}
  \item{X2Pp}{2-point field goal percentage}
  \item{eFGp}{effective field goal percentage: This statistic adjusts for the
    fact that a 3-point field goal is worth one more point that a 2-point field
    goal}
  \item{FT}{free throws per game}
  \item{FTA}{free throw attempts per game}
  \item{FTp}{free throw percentage}
  \item{ORB}{offensive rebounds per game}
  \item{DRB}{defensive rebounds per game}
  \item{TRB}{total rebounds per game}
  \item{AST}{assists per game}
  \item{STL}{steals per game}
  \item{BLK}{blocks per game}
  \item{TOV}{turnovers per game}
  \item{PF}{personal fouls per game}
  \item{PTS}{points per game}
}
}
\usage{
datasets
}
\description{
A dataset containing the players' 2022-2023 salary and 2021-2022 NBA
 performance data. Also, those players are not have accurate contract for next
 season yet. The variables are as follows:
}
\keyword{datasets}
