% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataTeam.2021.2022}
\alias{dataTeam.2021.2022}
\title{predict the championship}
\format{
a data frame with 30 rows and 41 columns
\describe{
 \item{Team}{the complete name of the team}
 \item{Team abbreviation}{the abbreviation name of the team}
 \item{PTS}{the average points that a team gets per game}
 \item{FGM}{the average field goals that a team makes per game}
 \item{FGA}{the average field goals that a team attempts per game}
 \item{FG\%}{the average field goal percentage of a team per game}
 \item{3pM}{the average three points that a team makes per game}
 \item{3PA}{the average three points that a team attempts per game}
 \item{3P\%}{the average three point percentage of a team per game}
 \item{FTM}{the average free throws that a team makes per game}
 \item{FTA}{the average free throws that a team attempts per game}
 \item{FT\%}{the average free throw percentage of a team per game}
 \item{OR}{the average offensive rebounds that a team gets per game}
 \item{DR}{the average defensive rebounds that a team gets per game}
 \item{REB}{the average total rebounds that a team gets per game}
 \item{AST}{the average assists that a team gets per game}
 \item{STL}{the average steals that a team gives per game}
 \item{BLK}{the average blocks that a team gives per game}
 \item{TO}{the average turnovers that a team has per game}
 \item{PF}{the average fouls that a team has per game}
 \item{OPTS}{the average points that a team's opponent gets per game}
 \item{OFGM}{the average field goals that a team's opponent makes per game}
 \item{OFGA}{the average field goals that a team's opponent attempts per game}
 \item{OFG\%}{the average field goal percentage of a team's opponent per game}
 \item{O3pM}{the average three points that a team's opponent makes per game}
 \item{O3PA}{the average three points that a team's opponent attempts per game}
 \item{O3P\%}{the average three point percentage of a team's opponent per game}
 \item{OFTM}{the average free throws that a team's opponent makes per game}
 \item{OFTA}{the average free throws that a team's opponent attempts per game}
 \item{OFT\%}{the average free throw percentage of a team's opponent per game}
 \item{OOR}{the average offensive rebounds that a team's opponent gets per game}
 \item{ODR}{the average defensive rebounds that a team's opponent gets per game}
 \item{OREB}{the average total rebounds that a team's opponent gets per game}
 \item{OAST}{the average assists that a team's opponent gets per game}
 \item{OSTL}{the average steals that a team's opponent gives per game}
 \item{OBLK}{the average blocks that a team's opponent gives per game}
 \item{OTO}{the average turnovers that a team's opponent has per game}
 \item{OPF}{the average fouls that a team's opponent has per game}
 \item{GP}{the game that each team plays in this season, it equals 82 for each team}
 \item{GW}{the game that each team wins in this season}
 \item{SW}{the series that each team wins in the playoff.
 For those teams who doesn't get into the playoff, it equals -1.
 For those teams who gets into the playoff, it equals the number of series that it wins.
 In order to get the championship, the team has to win 4 series.
 (-1 -- 4)}
}
}
\usage{
dataTeam.2021.2022
}
\description{
A data set containing data of the regular season 2021-2022 of all the teams,
along with the number of series each wins in the playoffs.
}
\keyword{datasets}
