% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-nba-player.R
\name{field_nba_player}
\alias{field_nba_player}
\title{Plot the field goal percentage (FGP) of NBA players for a given year and number of top players to display.}
\usage{
field_nba_player(index = 30, year = number, desc = T)
}
\arguments{
\item{index}{Integer, indicating the number of top players to display}

\item{year}{Integer, indicating the year of the NBA season to plot FGP for}

\item{desc}{Logical value, indicating whether to sort the players in descending order by FGP and games played (G)}
}
\value{
A plot of the top players with their FGP and G displayed
}
\description{
Plot the field goal percentage (FGP) of NBA players for a given year and number of top players to display.
}
\examples{
field_nba_player(index = 30, year = 2021, desc = TRUE)
}
