% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game.data}
\alias{game.data}
\title{analyze the correlation between different attributes to the final result of a game}
\format{
a data frame with 30 rows and 40 columns
\describe{
  \item{Team}{the complete name of the team}
  \item{Team abbreviation}{the abbreviation name of the team}
  \item{PTS}{the average points that a team gets per game}
  \item{FGM}{the average field goals that a team makes per game}
  \item{FGA}{the average field goals that a team attempts per game}
  \item{FG\%}{the average field goal percentage of a team per game}
  \item{3pM}{the average three points that a team makes per game}
  \item{3PA}{the average three points that a team attempts per game}
  \item{3P\%}{the average three point percentage of a team per game}
  \item{FTM}{the average free throws that a team makes per game}
  \item{FTA}{the average free throws that a team attempts per game}
  \item{FT\%}{the average free throw percentage of a team per game}
  \item{OR}{the average offensive rebounds that a team gets per game}
  \item{DR}{the average defensive rebounds that a team gets per game}
  \item{REB}{the average total rebounds that a team gets per game}
  \item{AST}{the average assists that a team gets per game}
  \item{STL}{the average steals that a team gives per game}
  \item{BLK}{the average blocks that a team gives per game}
  \item{TO}{the average turnovers that a team has per game}
  \item{PF}{the average fouls that a team has per game}
  \item{OPTS}{the average points that a team's opponent gets per game}
  \item{OFGM}{the average field goals that a team's opponent makes per game}
  \item{OFGA}{the average field goals that a team's opponent attempts per game}
  \item{OFG\%}{the average field goal percentage of a team's opponent per game}
  \item{O3pM}{the average three points that a team's opponent makes per game}
  \item{O3PA}{the average three points that a team's opponent attempts per game}
  \item{O3P\%}{the average three point percentage of a team's opponent per game}
  \item{OFTM}{the average free throws that a team's opponent makes per game}
  \item{OFTA}{the average free throws that a team's opponent attempts per game}
  \item{OFT\%}{the average free throw percentage of a team's opponent per game}
  \item{OOR}{the average offensive rebounds that a team's opponent gets per game}
  \item{ODR}{the average defensive rebounds that a team's opponent gets per game}
  \item{OREB}{the average total rebounds that a team's opponent gets per game}
  \item{OAST}{the average assists that a team's opponent gets per game}
  \item{OSTL}{the average steals that a team's opponent gives per game}
  \item{OBLK}{the average blocks that a team's opponent gives per game}
  \item{OTO}{the average turnovers that a team's opponent has per game}
  \item{OPF}{the average fouls that a team's opponent has per game}

}
}
\usage{
game.data
}
\description{
A data set containing different attributes of all the 30 teams in the league of the season 2022-2023
The variables are as follows.
}
\keyword{datasets}
