% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSW}
\alias{GSW}
\title{NBA Plays from 2022-2023 Season(GSW)}
\format{
A data frame with 40112 rows and 19 variables:
\describe{
  \item{game_id}{Game ID}
  \item{period}{Period of Play}
  \item{clock}{Game-clock at time of play}
  \item{home}{home team}
  \item{scoreHome}{Home team points at time of play}
  \item{away}{Away team}
  \item{scoreAway}{Away team points at time of play}
  \item{playerNameI}{Player making the play}
  \item{teamTricode}{Team of the player making the play}
  \item{description}{The description of the play}
  \item{actionType}{What kind of play}
  \item{subType}{A more detailed description of an actionType}
  \item{xLegacy}{The horizontal position of a player's shot}
  \item{yLegacy}{The vertical position of a player's shot}
  \item{shotDistance}{The distance from the basket of a field goal}
  \item{isFieldGoal}{Whether the play is or isn't a field goal}
  \item{shotVal}{The number of points that could be scored from a play}
  \item{scoreVal}{The number of points that were scored on a play}
  \item{location}{Which side of the court the play occurred}
}
}
\usage{
GSW
}
\description{
A dataset containing GSW competition performance on 2022-2023 Season. The
 variables are as follows:
}
\keyword{datasets}
