% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerStateAnalysis.R
\name{Kawhi.Leonard.state.analysis}
\alias{Kawhi.Leonard.state.analysis}
\title{analysis of state fluctuation of a superstar player: Kawhi Leonard}
\usage{
Kawhi.Leonard.state.analysis(
  data.category = c("MIN", "PTS", "REB", "FGM"),
  month.range = c("October", "November", "December", "January", "February", "March",
    "April")
)
}
\arguments{
\item{data.category}{the data category that we want to their changes, they are set as "MIN","PTS","REB","FGM", "REB" as the default
However, you can choose any data category in the data set.
They include "GP    W    L    MIN    PTS    FGM    FGA    FG\%    3PM    3PA    3P\%    FTM    FTA    FT\%    OREB    DREB    REB    AST    TOV    STL    BLK    PF    FP    DD2    TD3    BPM"}

\item{month.range}{the month range of Kawhi Leonard. They are required to be a subset of
("October","November","December","January","February","March","April") with the same order.}
}
\value{
the state change of the player, as a form of data frame,
column names are the month names, from October to April
row names are the average points, rebounds, assists, steals, blocks, turnovers of that player
}
\description{
analysis of state fluctuation of a superstar player: Kawhi Leonard
}
\examples{
library(NBADP)
Kawhi.Leonard.state.analysis(data.category=c("GP","STL","BLK","BPM"),
month.range=c("December","January","February","March"))
}
