% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lg}
\alias{lg}
\title{NBA League Average data in each year}
\format{
A data frame with 77 rows and 32 variables:
\describe{
  \item{Rk}{time ranking (1--77)}
  \item{Season}{the year of NBA season (1946-47--2022-23)}
  \item{Lg}{name of league (NBA)}
  \item{Age}{the average age of the league in each year (26.1--27.9)}
  \item{Ht}{average height of the league in each year (6-4--6-7)}
  \item{Wt}{average weight of the league in each year (195--223)}
  \item{G}{average number of games(192--1230)}
  \item{MP}{minutes played per game (240.5--242.2)}
  \item{FG}{field goals per game (25.9--45.9)}
  \item{FGA}{field goal attempts per game (75.4--109.4)}
  \item{3P}{3-point field goals per game (0.5--12.7)}
  \item{3PA}{3-point field goal attempts per game (2.0--35.2)}
  \item{FT}{free throws per game (15.9--28.6)}
  \item{FTA}{free throw attempts per game (21.7--38.3)}
  \item{ORB}{offensive rebounds per game (9.7--15.1)}
  \item{DRB}{defensive rebounds per game (28.4--34.8)}
  \item{TRB}{total rebounds per game (41.0--73.5)}
  \item{AST}{assists per game (6.9--26.3)}
  \item{STL}{steals per game (7.2--9.6)}
  \item{BLK}{blocks per game (4.3--5.6)}
  \item{TOV}{turnovers per game (13.8--20.8)}
  \item{PF}{personal fouls per game (19.3--28.8)}
  \item{PTS}{points per game (67.8--118.8)}
  \item{FG\%}{field goals percentage (0.279--0.492)}
  \item{3P\%}{3-point field goals percentage (0.238--0.367)}
  \item{FT\%}{free throw percentage (0.641--0.782)}
  \item{Pace}{pace factor: an estimate of possessions per 48 minutes (88.9--107.8)}
  \item{eFG\%}{effective field goal percentage: This statistic adjusts for the
    fact that a 3-point field goal is worth one more point that a 2-point field
    goal (0.279--0.545)}
  \item{TOV\%}{turnover percentage: an estimate of turnovers committed per 100
    plays (12.3--16.5)}
  \item{ORB\%}{offensive rebound percentage: an estimate of the percentage of
    available offensive rebounds a player grabbed while they were on the floor
    (22.2--33.5)}
  \item{FT/FGA}{free throws per field goal attempt (0.171--0.333)}
  \item{ORtg}{offensive rating: an estimate of points produced (players) or
    scored (teams) per 100 possessions (97.7--114.8)}
}
}
\usage{
lg
}
\description{
A dataset containing the league average level in many different variables.
 The variables are as follows:
}
\keyword{datasets}
