% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP_Prediction.R
\name{MVP.prediction}
\alias{MVP.prediction}
\title{predict the regular season MVP of the season 2022-2023, given the player's personal data of regular season.}
\usage{
MVP.prediction()
}
\value{
a data frame representing the rank of the MVP voting,
so that we know the first one will be the predicted player with the highest probability to be the MVP
}
\description{
predict the regular season MVP of the season 2022-2023, given the player's personal data of regular season.
}
\examples{
library(NBADP)
MVP.prediction()
}
