% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-win-rate.R
\name{nba_win_rate}
\alias{nba_win_rate}
\title{Predict NBA Team Win Rate}
\usage{
nba_win_rate(year = 2023, desc = T)
}
\arguments{
\item{year}{Integer, the year to predict the win rate for, defaults to 2023}

\item{desc}{Logical value, Whether to sort the output by descending predicted win rate}
}
\value{
A bar chart that shows the predicted win rate and the historic win rate for each team
}
\description{
Predict NBA Team Win Rate
}
\examples{
nba_win_rate(year = 2023)
}
