% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBADP.R
\docType{package}
\name{NBADP}
\alias{NBADP}
\title{NBADP: a package for analyzing the data in NBA}
\description{
The package gives three categories of functions :
prediction, regression and visualization
}
\section{prediction}{

predict the championship, MVP, salary of players, winning rate,
}

\section{regression}{

analyze the cost-performance ratio, the correlation between different attributes and the result of the game
}

\section{visualization}{

visualize Kawhi Leonard's state fluctuation, shot performance and distribution of players and teams
visualize field goal percentage
}

