% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PER.R
\name{PER}
\alias{PER}
\title{Player Efficiency Rate}
\usage{
PER(lgdata = lg, regulardata = X2021to2022data, yrs = "2021-22", player1, team)
}
\arguments{
\item{lgdata}{dataset, a dataset that contain the variables that have
league's basic information}

\item{regulardata}{dataset, a dataset that contain the variables that have basic
information of player's performance}

\item{yrs}{string, the seasons that we want to calculate}

\item{player1}{string, input the name of the player that we want to test}

\item{team}{string, input the team that the player on}
}
\value{
a number
}
\description{
We calculate player efficiency rating (PER) to boil down a player's performance.
}
\examples{
library(NBADP)
PER(lgdata = lg, regulardata = X2021to2022data, yrs = "2021-22", "LeBron James","LAL")
}
