% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerVisual.R
\name{player_visual}
\alias{player_visual}
\title{Visualize a player's shot percentage and distribution on a season}
\usage{
player_visual(teamdata = GSW, playerName, teamName)
}
\arguments{
\item{teamdata}{dataset, a dataset that contain the team's all regular match data in 2022-2023 season}

\item{playerName}{string, input the player's name}

\item{teamName}{string, input the team that the player on}
}
\value{
The shot plot of the player in the whole regular season
}
\description{
Visualize a player's shot percentage and distribution on a season
}
\examples{
library(NBADP)
player_visual(teamdata = GSW,"S. Curry","GSW")
}
