% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSalary.R
\name{pred_salary}
\alias{pred_salary}
\title{Make a salary forecast for 2023-2024}
\usage{
pred_salary(
  turn_over,
  points_per_game,
  two_point_field_goal_attempts_per_game,
  free_throws_per_game,
  assists_per_game,
  games_started,
  value_over_replacement_player,
  defensive_rebounds_per_game,
  minutes_played_per_game,
  three_point_field_goal_attempts_per_game,
  usage_percentage,
  offensive_box_plusorminus,
  defensive_win_shares
)
}
\arguments{
\item{turn_over}{numeric, turnovers per game (TOV).}

\item{points_per_game}{numeric, points per game (PTS).}

\item{two_point_field_goal_attempts_per_game}{numeric, (2PA).}

\item{free_throws_per_game}{numeric, free throws per game (FT).}

\item{assists_per_game}{numeric, assists per game (AST).}

\item{games_started}{numeric, games started (GS).}

\item{value_over_replacement_player}{numeric, value over replacement player (VORP).}

\item{defensive_rebounds_per_game}{numeric, defensive rebounds per game (DRB).}

\item{minutes_played_per_game}{numeric, minutes played per game (MP).}

\item{three_point_field_goal_attempts_per_game}{numeric, three point field goal attempts per game (3PA).}

\item{usage_percentage}{numeric, usage percentage (USG\%).}

\item{offensive_box_plusorminus}{numeric, offensive box plus or minus (OBPM).}

\item{defensive_win_shares}{numeric, defensive win shares (DWS).}
}
\value{
according to the player's performance in 2023-2024 to predict 2023-2024 salary
}
\description{
Make a salary forecast for 2023-2024
}
\examples{
library(NBADP)
pred_salary(0.6, 3.9, 2.1, 0.4, 1.4, 6, 0, 0.8, 13.4, 1.4, 13.4, -2.9, 0.6)

}
