% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shot}
\alias{shot}
\title{NBA Shots Dataset(20211205)}
\format{
A data frame with 690 rows and 16 variables:
\describe{
  \item{ ...1}{variable ranking}
  \item{Unnamed: 0}{variable ranking}
  \item{Unnamed: 0.1}{variable ranking}
  \item{match_id}{game identifier. Unique for each matchup}
  \item{shotX}{x coordinate of the shot position}
  \item{shotY}{y coordinate of the shot position}
  \item{quarter}{quarter of the game}
  \item{time_remaining}{time remaining in the quarter}
  \item{player}{name of the player who took the shot}
  \item{team}{3 letter abbreviation of the attacking team}
  \item{made}{whether the shot resulted in a made basket or not}
  \item{shot_type}{2-pointer or 3-pointer}
  \item{distance}{Distance (in ft) of shot from the basket. Dunks count as
   shots made from 0 distance}
  \item{score}{game score after the shot}
  \item{opp}{3 letter abbreviation of the defending team}
  \item{status}{situation of score (tied,trials,leads)}
}
}
\usage{
shot
}
\description{
A dataset containing the competition performance on date 2021/12/05.
 The variables are as follows:
}
\keyword{datasets}
