% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotVisual.R
\name{shot_visual}
\alias{shot_visual}
\title{Visualize the 2 PT percentage and 3 PT percentage of all the players}
\usage{
shot_visual(data1 = shot, team1, opponent1)
}
\arguments{
\item{data1}{dataset, dataset that contains all match data of the day}

\item{team1}{string, input the team that we want to visualize}

\item{opponent1}{string, input the non-home team of this match}
}
\value{
plot, containing the situation of whether the basketball shot is in
or out
}
\description{
The plot that show the corresponding match shot result
}
\examples{
library(NBADP)
shot_visual(data1 = shot, team1 = "NOP",opponent1 = "NOP")

}
