% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.data.raw}
\alias{state.data.raw}
\title{Analyze the state change of Kawhi Leonard
the state fluctuation of Kawhi Leonard each month, from 2022 October to 2023 April}
\format{
a data frame with 7 rows and 30 variables:
\describe{
  \item{Player}{the name of the player(Kawhi Leonard)}
  \item{Month}{the month that he plays(October--April)}
  \item{TEAM}{the team that Kawhi Leonard plays for(LAC)}
  \item{AGE}{the age of Kawhi Leonard(31)}
  \item{GP}{the number of games that Kawhi Leonard plays}
  \item{W}{the number of games that Kawhi Leonard wins}
  \item{L}{the number of games that Kawhi Leonard loses}
  \item{MIN}{the average time that Kawhi Leonard plays per game}
  \item{PTS}{the average points that Kawhi Leonard gets per game}
  \item{FGM}{the average field goals that Kawhi Leonard made per game}
  \item{FGA}{the average field goals that Kawhi Leonard attempted per game}
  \item{FG\%}{the average field goal percentage of Kawhi Leonard per game}
  \item{3pM}{the average three point field goals that Kawhi Leonard made per game}
  \item{3pA}{the average three point field goals that Kawhi Leonard attempted per game}
  \item{3p\%}{the average three point field percentage percentage of Kawhi Leonard per game}
  \item{FTM}{the average free throws that Kawhi Leonard made per game}
  \item{FTA}{the average free throws that Kawhi Leonard attempted per game}
  \item{FT\%}{the average free throw percentage of Kawhi Leonard per game}
  \item{OREB}{the average offensive rebounds that Kawhi Leonard gets per game}
  \item{DREB}{the average defensive rebounds that Kawhi Leonard gets per game}
  \item{REB}{the average total rebounds that Kawhi Leonard gets per game}
  \item{AST}{the average assists that Kawhi Leonard gets per game}
  \item{TOV}{the average turnovers that Kawhi Leonard makes per game}
  \item{STL}{the average steals that Kawhi Leonard gets per game}
  \item{BLK}{the average blocks that Kawhi Leonard gives per game}
  \item{PF}{the average personal fouls that Kawhi Leonard has per game}
  \item{FP}{the average fantasy points that Kawhi Leonard has per game}
  \item{DD2}{the number of double doubles that Kawhi Leonard has in the month}
  \item{TD3}{the number of triple doubles that Kawhi Leonard has in the month}
  \item{BPM}{the value of Plus-Minus of Kawhi Leonard per game}
}
}
\usage{
state.data.raw
}
\description{
A data set containing different attributes of Kawhi Leonard in those months.
The variables are as follows.
}
\keyword{datasets}
