% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X2021to2022data}
\alias{X2021to2022data}
\title{2021-2022 NBA Player Stats: Per Game}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 812 rows and 31 columns.
}
\usage{
X2021to2022data
}
\description{
A dataset containing each player's details performance data in the 2021-2022
 NBA regular season. The variables are as follows:
}
\details{
@format A data frame with 812 rows and 31 variables:
 \describe{
   \item{Rk}{player ranking (1--605)}
   \item{Player}{player name}
   \item{Pos}{position}
   \item{Age}{player's age (19--41)}
   \item{Tm}{3 letter abbreviation of the team}
   \item{G}{number of games that the player join in (1-82)}
   \item{GS}{game started (0--82)}
   \item{MP}{minutes played per game(1--2854)}
   \item{FG}{field goals per game (0--774)}
   \item{FGA}{field goal attempts per game (0--1564)}
   \item{FG\%}{field goal percentage (0--1)}
   \item{3P}{3-point field goals per game (0--285)}
   \item{3PA}{3-point field goal attempts per game (0--750)}
   \item{3P\%}{3-point field goal percentage (0--1)}
   \item{2P}{2-point field goals per game (0--724)}
   \item{2PA}{2-point field goal attempts per game (0--1393)}
   \item{2P\%}{2-point field goal percentage (0--1)}
   \item{eFG\%}{effective field goal percentage: This statistic adjusts for the
    fact that a 3-point field goal is worth one more point that a 2-point field
    goal (0--1)}
   \item{FT}{free throws per game (0--654)}
   \item{FTA}{free throw attempts per game (0--803)}
   \item{FT\%}{free throw percentage (0--1)}
   \item{ORB}{offensive rebounds per game (0--349)}
   \item{DRB}{defensive rebounds per game (0--813)}
   \item{TRB}{total rebounds per game (0--1019)}
   \item{AST}{assists per game (0--737)}
   \item{STL}{steals per game (0--138)}
   \item{BLK}{blocks per game (0--177)}
   \item{TOV}{turnovers per game (0--303)}
   \item{PF}{personal fouls per game (0--286)}
   \item{PTS}{points per game (0--2155)}
   \item{Player-additional}{additional information}
}
}
\keyword{datasets}
